Software - Fixes for Linux and updated stationery

+++ Code_Design/DigitalLabEquipment/Project_Headers/gpio.h
@@ -78,10 +78,10 @@ namespace USBDM {
  * @tparam bitNum          Bit number within PORT/GPIO
  * @tparam polarity        Polarity of pin. Either ActiveHigh or ActiveLow
  */
-template<uint32_t clockInfo, uint32_t portAddress, IRQn_Type irqNum, uint32_t gpioAddress, unsigned bitNum, Polarity polarity>
+template<uint32_t clockInfo, uint32_t portAddress, IRQn_Type irqNum, uint32_t gpioAddress, int bitNum, Polarity polarity>
 class GpioBase_T {
 
-   static_assert((bitNum<=31), "Illegal bit number in Gpio");
+   static_assert((static_cast<unsigned>(bitNum)<=31), "Illegal bit number in Gpio");
 
 private:
    /**
